Refactor and add tests on celery tasks
