+++ simulator_api/celery_instance/celery.py
@@ -1,0 +1,10 @@
+from celery import Celery
+
+celery_instance = Celery(
+    'entrypoint',
+    broker='pyamqp://guest:guest@localhost:5672//',
+    backend='db+sqlite:////opt/mov.ai/app/celery_data/results.sqlite3'
+)
+celery_instance.conf.broker_connection_retry_on_startup = True
+celery_instance.conf.worker_hijack_root_logger = False
+celery_instance.conf.task_track_started = True
