+++ tests/test_commands/test_communication_test.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest import mock
 
-from simulator_api.commands.communication_test import CommunicationTest
+from simulator_api.commands.communication_test import CommunicationTest, communication_test
 
 mock_celery_task_obj = mock.MagicMock()
 mock_celery_task_obj.id = 12345
@@ -24,7 +24,7 @@ class TestCommandCommunicationTest(unittest.TestCase):
         self.assertEqual(response.content, {'task_id': 12345})
 
     @mock.patch('simulator_api.commands.communication_test.communication_test.AsyncResult')    
-    def test_post_execute_communication_test(self, mock_communincation_test_async_result):
+    def test_get_execute_communication_test(self, mock_communincation_test_async_result):
 
         mock_communincation_test_async_result.return_value = mock_celery_task_obj
 
@@ -35,3 +35,13 @@ class TestCommandCommunicationTest(unittest.TestCase):
         mock_communincation_test_async_result.assert_called_once()
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.content, {'status': "SUCCESS", 'info': "hello"})
+
+    @mock.patch('simulator_api.commands.communication_test.communication_test.update_state')
+    def test_communication_test(self, mock_communincation_test_update):        
+
+        result = communication_test()
+        # Expected results without ignition installed
+        self.assertEqual(result['status'],"ERROR")
+        self.assertEqual(len(result['checklist']),6)
+        for task_result in result['checklist']:
+            self.assertEqual(task_result['status'],"ERROR")
